+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Test and beta releases will have experimental functions enabled by default.
 
 **Next release**
 
+* Show restrictions of system applications, even when disabled
+
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 1.99.11 TEST**