+++ CHANGELOG.md
@@ -15,10 +15,12 @@ Test and beta releases will have experimental functions enabled by default.
 
 **Next release**
 
+[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
+
+**Version 1.99.12 TEST**
+
 * Show restrictions of system applications, even when disabled
 * Moved database back due to permission problems ([issue](https://github.com/M66B/XPrivacy/issues/1129))
-
-[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 1.99.11 TEST**